perf: use paginated batches on home page
