+++ web-registry/src/hooks/batches/useBatchesWithSupply.tsx
