+++ web-registry/src/pages/Dashboard/MyCreditBatches/MyCreditBatches.tsx
@@ -1,24 +1,17 @@
-import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useTheme } from '@mui/styles';
 import { Box } from '@mui/system';
-import { QueryBatchesByIssuerResponse } from '@regen-network/api/lib/generated/regen/ecocredit/v1/query';
 
 import OutlinedButton from 'web-components/lib/components/buttons/OutlinedButton';
 import EmptyState from 'web-components/lib/components/empty-state';
 import { CreditBatchLightIcon } from 'web-components/lib/components/icons/CreditBatchLightIcon';
 import PlusIcon from 'web-components/lib/components/icons/PlusIcon';
-import {
-  DEFAULT_ROWS_PER_PAGE,
-  OnActionTableChangeParams,
-} from 'web-components/lib/components/table/ActionsTable';
 import { Title } from 'web-components/lib/components/typography';
 
 import WithLoader from 'components/atoms/WithLoader';
 import { CreditBatches } from 'components/organisms';
-import { useEcocreditQuery } from 'hooks';
+import { usePaginatedBatchesByIssuer } from 'hooks/batches/usePaginatedBatchesByIssuer';
 
-import { useBatchesWithSupply } from './hooks/useBatchesWithSupply';
 import { NO_CREDIT_BATCHES_MESSAGE } from './MyCreditBatches.constants';
 
 type MyCreditBatchesProps = {
@@ -29,20 +22,8 @@ export const MyCreditBatches = ({
   address,
 }: MyCreditBatchesProps): JSX.Element => {
   const theme = useTheme();
-  const [paginationParams, setPaginationParams] =
-    useState<OnActionTableChangeParams>({
-      page: 0,
-      rowsPerPage: DEFAULT_ROWS_PER_PAGE,
-      offset: 0,
-    });
-  const batchesResponse = useEcocreditQuery<QueryBatchesByIssuerResponse>({
-    params: { issuer: address },
-    query: 'batchesByIssuer',
-  });
-  const batchesWithSupply = useBatchesWithSupply({
-    batches: batchesResponse?.data?.batches,
-    paginationParams,
-  });
+  const { batchesWithSupply, setPaginationParams } =
+    usePaginatedBatchesByIssuer({ address });
   const hasNoBatches = batchesWithSupply && batchesWithSupply?.length === 0;
 
   return (
