+++ web-registry/src/pages/Home/Home.tsx
@@ -2,10 +2,7 @@ import React, { useEffect, useMemo, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { Box, CardMedia, useMediaQuery } from '@mui/material';
 import { useTheme } from '@mui/styles';
-import {
-  QueryBatchesResponse,
-  QueryProjectsResponse,
-} from '@regen-network/api/lib/generated/regen/ecocredit/v1/query';
+import { QueryProjectsResponse } from '@regen-network/api/lib/generated/regen/ecocredit/v1/query';
 
 import { BlockContent } from 'web-components/lib/components/block-content';
 import ContainedButton from 'web-components/lib/components/buttons/ContainedButton';
@@ -13,13 +10,8 @@ import ProjectCard from 'web-components/lib/components/cards/ProjectCard';
 import { Loading } from 'web-components/lib/components/loading';
 import Modal from 'web-components/lib/components/modal';
 import Section from 'web-components/lib/components/section';
-import {
-  DEFAULT_ROWS_PER_PAGE,
-  OnActionTableChangeParams,
-} from 'web-components/lib/components/table/ActionsTable';
 import { Body, Title } from 'web-components/lib/components/typography';
 
-import { useBatchesWithSupply } from 'pages/Dashboard/MyCreditBatches/hooks/useBatchesWithSupply';
 import { useProjectsSellOrders } from 'pages/Projects/hooks/useProjectsSellOrders';
 import { useSortProjects } from 'pages/Projects/hooks/useSortProjects';
 import {
@@ -28,6 +20,7 @@ import {
 } from 'pages/Projects/Projects.config';
 import WithLoader from 'components/atoms/WithLoader';
 import { useEcocreditQuery } from 'hooks';
+import { usePaginatedBatches } from 'hooks/batches/usePaginatedBatches';
 import { useQuerySellOrders } from 'hooks/useQuerySellOrders';
 
 import topographyImg from '../../assets/background-contour-1.jpg';
@@ -82,21 +75,7 @@ const Home: React.FC = () => {
     sort: PROJECTS_SORT,
   });
 
-  // Credit batches fetching
-  const [paginationParams, setPaginationParams] =
-    useState<OnActionTableChangeParams>({
-      page: 0,
-      rowsPerPage: DEFAULT_ROWS_PER_PAGE,
-      offset: 0,
-    });
-  const batchesResponse = useEcocreditQuery<QueryBatchesResponse>({
-    params: {},
-    query: 'batches',
-  });
-  const batchesWithSupply = useBatchesWithSupply({
-    batches: batchesResponse?.data?.batches,
-    paginationParams,
-  });
+  const { batchesWithSupply, setPaginationParams } = usePaginatedBatches();
 
   const isMobile = useMediaQuery(theme.breakpoints.down('md'));
